+++ Dockerfile.test
@@ -1,0 +1,5 @@
+FROM golang:1.22.1-alpine
+
+WORKDIR /go/src/target
+
+CMD CGO_ENABLED=0 go test -v --tags=integration ./...
