+++ handler/handler.go
@@ -1,0 +1,109 @@
+package handler
+
+import (
+	"database/sql"
+	"net/http"
+
+	"github.com/labstack/echo/v4"
+	"github.com/lib/pq"
+)
+
+type handler struct {
+	DB *sql.DB
+}
+
+func NewApplication(db *sql.DB) *handler {
+	return &handler{db}
+}
+
+func (h *handler) Greeting(c echo.Context) error {
+	return c.String(http.StatusOK, "Hello, World!")
+}
+
+type Expense struct {
+	ID     int      `json:"id"`
+	Title  string   `json:"title"`
+	Amount float32  `json:"amount"`
+	Note   string   `json:"note"`
+	Tags   []string `json:"tags"`
+}
+
+type Err struct {
+	Message string `json:"message"`
+}
+
+func (h *handler) CreateExpenseHandler(c echo.Context) error {
+	e := Expense{}
+	err := c.Bind(&e)
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, Err{Message: err.Error()})
+	}
+
+	row := h.DB.QueryRow("INSERT INTO expenses (title, amount, note, tags) values ($1, $2, $3, $4) RETURNING id", e.Title, e.Amount, e.Note, pq.Array(e.Tags))
+	err = row.Scan(&e.ID)
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, Err{Message: err.Error()})
+	}
+	return c.JSON(http.StatusCreated, e)
+}
+
+func (h *handler) GetExpenseHandler(c echo.Context) error {
+	id := c.Param("id")
+	stmt, err := h.DB.Prepare("SELECT id, title, amount, note, tags FROM expenses WHERE id = $1")
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, Err{Message: "can't prepare query expense statment:" + err.Error()})
+	}
+
+	row := stmt.QueryRow(id)
+	e := Expense{}
+	err = row.Scan(&e.ID, &e.Title, &e.Amount, &e.Note, pq.Array(&e.Tags))
+	switch err {
+	case sql.ErrNoRows:
+		return c.JSON(http.StatusNotFound, Err{Message: "expense not found"})
+	case nil:
+		return c.JSON(http.StatusOK, e)
+	default:
+		return c.JSON(http.StatusInternalServerError, Err{Message: "can't scan expense:" + err.Error()})
+	}
+}
+
+func (h *handler) UpdateExpenseHandler(c echo.Context) error {
+	id := c.Param("id")
+	e := Expense{}
+	err := c.Bind(&e)
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, Err{Message: err.Error()})
+	}
+
+	row := h.DB.QueryRow("UPDATE expenses SET title = $2, amount = $3, note = $4, tags = $5 WHERE id = $1 RETURNING id", id, e.Title, e.Amount, e.Note, pq.Array(e.Tags))
+	err = row.Scan(&e.ID)
+
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, Err{Message: err.Error()})
+	}
+	return c.JSON(http.StatusCreated, e)
+}
+
+func (h *handler) GetExpensesHandler(c echo.Context) error {
+	stmt, err := h.DB.Prepare("SELECT id, title, amount, note, tags FROM expenses")
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, Err{Message: "can't prepare query all expenses statement:" + err.Error()})
+	}
+
+	rows, err := stmt.Query()
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, Err{Message: "can't query all expenses:" + err.Error()})
+	}
+
+	expenses := []Expense{}
+
+	for rows.Next() {
+		e := Expense{}
+		err := rows.Scan(&e.ID, &e.Title, &e.Amount, &e.Note, pq.Array(&e.Tags))
+		if err != nil {
+			return c.JSON(http.StatusInternalServerError, Err{Message: "can't scan expense"})
+		}
+		expenses = append(expenses, e)
+	}
+	return c.JSON(http.StatusOK, expenses)
+}
