+++ server.go
@@ -12,96 +12,8 @@ import (
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
 
-	"github.com/lib/pq"
+	"github.com/openmymai/assessment/handler"
 )
-
-type Expense struct {
-	ID     int      `json:"id"`
-	Title  string   `json:"title"`
-	Amount float32  `json:"amount"`
-	Note   string   `json:"note"`
-	Tags   []string `json:"tags"`
-}
-
-type Err struct {
-	Message string `json:"message"`
-}
-
-func createExpenseHandler(c echo.Context) error {
-	e := Expense{}
-	err := c.Bind(&e)
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, Err{Message: err.Error()})
-	}
-
-	row := db.QueryRow("INSERT INTO expenses (title, amount, note, tags) values ($1, $2, $3, $4) RETURNING id", e.Title, e.Amount, e.Note, pq.Array(e.Tags))
-	err = row.Scan(&e.ID)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, Err{Message: err.Error()})
-	}
-	return c.JSON(http.StatusCreated, e)
-}
-
-func getExpenseHandler(c echo.Context) error {
-	id := c.Param("id")
-	stmt, err := db.Prepare("SELECT id, title, amount, note, tags FROM expenses WHERE id = $1")
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, Err{Message: "can't prepare query expense statment:" + err.Error()})
-	}
-
-	row := stmt.QueryRow(id)
-	e := Expense{}
-	err = row.Scan(&e.ID, &e.Title, &e.Amount, &e.Note, pq.Array(&e.Tags))
-	switch err {
-	case sql.ErrNoRows:
-		return c.JSON(http.StatusNotFound, Err{Message: "expense not found"})
-	case nil:
-		return c.JSON(http.StatusOK, e)
-	default:
-		return c.JSON(http.StatusInternalServerError, Err{Message: "can't scan expense:" + err.Error()})
-	}
-}
-
-func updateExpenseHandler(c echo.Context) error {
-	id := c.Param("id")
-	e := Expense{}
-	err := c.Bind(&e)
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, Err{Message: err.Error()})
-	}
-
-	row := db.QueryRow("UPDATE expenses SET title = $2, amount = $3, note = $4, tags = $5 WHERE id = $1 RETURNING id", id, e.Title, e.Amount, e.Note, pq.Array(e.Tags))
-	err = row.Scan(&e.ID)
-
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, Err{Message: err.Error()})
-	}
-	return c.JSON(http.StatusCreated, e)
-}
-
-func getExpensesHandler(c echo.Context) error {
-	stmt, err := db.Prepare("SELECT id, title, amount, note, tags FROM expenses")
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, Err{Message: "can't prepare query all expenses statement:" + err.Error()})
-	}
-
-	rows, err := stmt.Query()
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, Err{Message: "can't query all expenses:" + err.Error()})
-	}
-
-	expenses := []Expense{}
-
-	for rows.Next() {
-		e := Expense{}
-		err := rows.Scan(&e.ID, &e.Title, &e.Amount, &e.Note, pq.Array(&e.Tags))
-		if err != nil {
-			return c.JSON(http.StatusInternalServerError, Err{Message: "can't scan expense"})
-		}
-		expenses = append(expenses, e)
-	}
-	return c.JSON(http.StatusOK, expenses)
-}
 
 var db *sql.DB
 
@@ -124,7 +36,11 @@ func main() {
 		log.Fatal("can't create table", err)
 	}
 
+	h := handler.NewApplication(db)
+
 	e := echo.New()
+
+	e.GET("/", h.Greeting)
 
 	e.Use(middleware.BasicAuth(func(username, password string, c echo.Context) (bool, error) {
 		if username == "apidesign" || password == "45678" {
@@ -136,10 +52,10 @@ func main() {
 	e.Use(middleware.Logger())
 	e.Use(middleware.Recover())
 
-	e.POST("/expenses", createExpenseHandler)
-	e.GET("/expenses/:id", getExpenseHandler)
-	e.PUT("/expenses/:id", updateExpenseHandler)
-	e.GET("/expenses", getExpensesHandler)
+	e.POST("/expenses", h.CreateExpenseHandler)
+	e.GET("/expenses/:id", h.GetExpenseHandler)
+	e.PUT("/expenses/:id", h.UpdateExpenseHandler)
+	e.GET("/expenses", h.GetExpensesHandler)
 
 	go func() {
 		if err := e.Start(os.Getenv("PORT")); err != nil && err != http.ErrServerClosed {
